refactor: Move endpoint to ReviewController
